Account totals depends on categories
If the category of the transactions is either a credit payment or is
income then it is assumed to be positive. so it is added to the total.
Otherwise, the number is subtracted from the total. Added credit payment
as a "income" for credit accounts. Making NO MORE CHANGES  to traction
categories..