Added account balances and switching accounts bug
- Accounts have balances!
- When switching accounts, the datasources current value does not update
once something new has been selected in the list. This resulted in
transactions being displayed for the wrong account and once that account
was viewed, the transactions would also move to the previous account
selected. Changed this event to check if the actual current value of the
datasource has been changed to consistely get correct data.